backend/run_production.py
from waitress import serve
from main import create_app
import os
from dotenv import load_dotenv

load_dotenv()

app = create_app()

if __name__ == "__main__":
    port = int(os.getenv("PORT", 5001))
    print(f"Starting production server on port {port}...")
    serve(app, host="0.0.0.0", port=port, threads=8)
