backend/requirements.txt
Flask
Flask-RESTful
Flask-JWT-Extended
flask-mongoengine-3
pymongo
flask-pymongo
dnspython
Flask-CORS
python-dotenv
Werkzeug
reportlab==4.0.7
xhtml2pdf==0.2.16
num2words==0.5.13
waitress
mongoengine
flask-limiter
sentry-sdk
prometheus_client
pytest
mongomock